requirements.txt
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
fastapi==0.109.1
python-dotenv==0.21.0
uvicorn==0.20.0
sqlalchemy==2.0.30
psycopg2==2.9.9
alembic==1.8.1
apscheduler==3.9.0
requests==2.32.2
jupyter==1.0.0
seaborn==0.13.2
mlflow==2.14.2
optuna==4.0.0
joblib==1.4.2
gql==3.5.0
requests-toolbelt==1.0.0
pyyaml>=6.0